requirements.txt
requests
IMDbPY